requirements.txt
Python  = 3.9.0
APScheduler==3.7.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
forex-python==1.5
greenlet==1.0.0
idna==2.10
kaleido==0.2.1
lxml==4.6.3
multitasking==0.0.9
numpy==1.19.3
pandas==1.1.5
pandas-datareader==0.9.0
plotly==4.14.3
psutil==5.8.0
pyTelegramBotAPI==3.7.7
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
retrying==1.3.3
simplejson==3.17.2
six==1.15.0
soupsieve==2.2.1
SQLAlchemy==1.4.3
tzlocal==2.1
urllib3==1.26.4
yfinance==0.1.59
